Fix AccessorUtility warning

+++ CesiumGltf/include/CesiumGltf/AccessorUtility.h
@@ -201,7 +201,7 @@ struct IndicesForFaceFromAccessor {
     } else {
       for (int64_t i = 0; i < 3; i++) {
         int64_t vertexIndex = firstVertex + i;
-        result[i] = vertexIndex < vertexCount ? vertexIndex : -1;
+        result[static_cast<size_t>(i)] = vertexIndex < vertexCount ? vertexIndex : -1;
       }
     }
